Move modal related helpers from general to modal specific file
This makes it easier to get an overview of modal related code.
