+++ src/apps/dashboard/modal/ReservationsGroupModal.tsx
@@ -4,12 +4,12 @@ import { useText } from "../../../core/utils/text";
 import GroupModalContent from "../../../components/GroupModal/GroupModalContent";
 import { Button } from "../../../components/Buttons/Button";
 import SimpleModalHeader from "../../../components/GroupModal/SimpleModalHeader";
-import { getModalIds } from "../../../core/utils/helpers/general";
 import GroupModalReservationsList from "../../../components/GroupModal/GroupModalReservationsList";
 import { ReservationType } from "../../../core/utils/types/reservation-type";
 import StatusCircleModalHeader from "../../../components/GroupModal/StatusCircleModalHeader";
 import StatusCircle from "../../loan-list/materials/utils/status-circle";
 import useReservations from "../../../core/utils/useReservations";
+import { getModalIds } from "../../../core/utils/helpers/modal-helpers";
 
 interface ReservationGroupModalProps {
   pageSize: number;
