+++ src/apps/loan-list/list/loan-list.tsx
@@ -3,9 +3,7 @@ import { useSelector } from "react-redux";
 import dayjs from "dayjs";
 import {
   getAmountOfRenewableLoans,
-  getModalIds,
   getScrollClass,
-  constructModalId
 } from "../../../core/utils/helpers/general";
 import { getUrlQueryParam } from "../../../core/utils/helpers/url";
 import { useText } from "../../../core/utils/text";
@@ -14,7 +12,7 @@ import {
   ModalIdsProps
 } from "../../../core/utils/modal";
 import List from "./list";
-import { LoanType } from "../../../core/utils/types/loan-type";
+import { isLoanType, LoanType } from "../../../core/utils/types/loan-type";
 import { ListView } from "../../../core/utils/types/list-view";
 import EmptyList from "../../../components/empty-list/empty-list";
 import ToggleListViewButtons from "./ToggleListViewButtons";
@@ -24,19 +22,18 @@ import {
   removeLoansWithDuplicateDueDate,
   getFromListByKey
 } from "../utils/helpers";
-import MaterialDetails, {
-  constructMaterialDetailsModalId
-} from "../modal/material-details";
+import MaterialDetails from "../modal/material-details";
 import MaterialDetailsModal, {
   loanDetailsModalId
 } from "../modal/material-details-modal";
 import {
   getDetailsModalId,
   containsDueDateModalQueryParam,
-  dateFromDueDateModalQueryParam
+  dateFromDueDateModalQueryParam,
+  constructModalId,
+  getModalIds
 } from "../../../core/utils/helpers/modal-helpers";
 import LoansGroupModal from "../../../components/GroupModal/LoansGroupModal";
-import { ListType } from "../../../core/utils/types/list-type";
 import SimpleModalHeader from "../../../components/GroupModal/SimpleModalHeader";
 import StatusCircleModalHeader from "../../../components/GroupModal/StatusCircleModalHeader";
 import StatusCircle from "../materials/utils/status-circle";
@@ -54,10 +51,9 @@ const LoanList: FC<LoanListProps> = ({ pageSize }) => {
   const { loanDetails, allLoansId, dueDateModal, acceptModal } = getModalIds();
   const t = useText();
   const [view, setView] = useState<ListView>("list");
-  const [modalLoan, setModalLoan] = useState<ListType | null>(null);
   const [dueDate, setDueDate] = useState<string | null>(null);
   const [accepted, setAccepted] = useState<boolean>(false);
-  const [modalDetailsId, setModalDetailsId] = useState<string | null>(null);
+  const [modalLoan, setModalLoan] = useState<LoanType | null>(null);
   const {
     fbs: {
       loans: loansPhysical,
@@ -66,33 +62,13 @@ const LoanList: FC<LoanListProps> = ({ pageSize }) => {
     publizon: { loans: loansDigital }
   } = useLoans();
 
-  useEffect(() => {
-    // If modalLoan is already set it should not be set again, because it will cause an infinite loop
-    if (modalLoan) {
-      return;
-    }
-    let loanForModal = null;
-    if (loansPhysical && modalDetailsId) {
-      loanForModal = getFromListByKey(loansPhysical, "loanId", modalDetailsId);
-    }
-    if (loanForModal?.length === 0 && loansDigital && modalDetailsId) {
-      loanForModal = getFromListByKey(
-        loansDigital,
-        "identifier",
-        modalDetailsId
-      );
-    }
-    if (loanForModal && loanForModal.length > 0) {
-      setModalLoan(loanForModal[0]);
-    }
-  }, [modalDetailsId, modalLoan, loansPhysical, loansDigital]);
-
   const openAcceptModal = useCallback(() => {
     open(`${acceptModal}`);
   }, [acceptModal, open]);
 
   const openLoanDetailsModal = useCallback(
     (loan: LoanType) => {
+      setModalLoan(loan);
       open(loanDetailsModalId(loan));
     },
     [open]
@@ -116,12 +92,20 @@ const LoanList: FC<LoanListProps> = ({ pageSize }) => {
     // if there is a loan details query param, loan details modal should be opened
     const loanDetailsString = loanDetails as string;
     if (modalUrlParam && modalUrlParam.includes(loanDetails as string)) {
-      const loanDetailsModalId = getDetailsModalId(
+      const loanIdFromModalId = getDetailsModalId(
         modalUrlParam,
         loanDetailsString
       );
-      if (loanDetailsModalId) {
-        setModalDetailsId(loanDetailsModalId);
+      if (loanIdFromModalId && loansPhysical) {
+        const loans = getFromListByKey(
+          loansPhysical,
+          "loanId",
+          loanIdFromModalId
+        );
+        const loan = loans.filter(isLoanType).at(0);
+        if (loan) {
+          setModalLoan(loan);
+        }
       }
     }
 
@@ -130,7 +114,7 @@ const LoanList: FC<LoanListProps> = ({ pageSize }) => {
       const dateFromQueryParam = dateFromDueDateModalQueryParam(modalUrlParam);
       setDueDate(dateFromQueryParam);
     }
-  }, [allLoansId, loanDetails, openDueDateModal]);
+  }, [loansPhysical, loanDetails, openDueDateModal]);
 
   const listContainsLoans =
     (Array.isArray(loansPhysical) && loansPhysical.length > 0) ||
@@ -206,16 +190,11 @@ const LoanList: FC<LoanListProps> = ({ pageSize }) => {
       which is necessary to comply with WCAG (so the screen readers cannot "catch" focusable html
       elements below the modal) */}
       {modalLoan && (
-        <MaterialDetailsModal
-          modalId={constructMaterialDetailsModalId(loanDetails, modalDetailsId)}
-        >
+        <MaterialDetailsModal modalId={loanDetailsModalId(modalLoan)}>
           <MaterialDetails
             item={modalLoan}
             loan={modalLoan as LoanType}
-            modalId={constructMaterialDetailsModalId(
-              loanDetails,
-              modalDetailsId
-            )}
+            modalId={loanDetailsModalId(modalLoan)}
           />
         </MaterialDetailsModal>
       )}
