+++ src/apps/reservation-list/list/reservation-pause-toggler.tsx
@@ -1,10 +1,10 @@
 import React, { useState, FC, useEffect } from "react";
 import ReservationsIcon from "@danskernesdigitalebibliotek/dpl-design-system/build/icons/collection/Reservations.svg";
-import { getModalIds } from "../../../core/utils/helpers/general";
 import { useModalButtonHandler } from "../../../core/utils/modal";
 import { useText } from "../../../core/utils/text";
 import { PatronV5 } from "../../../core/fbs/model";
 import { formatDate } from "../../../core/utils/helpers/date";
+import { getModalIds } from "../../../core/utils/helpers/modal-helpers";
 
 interface ReservationPauseTogglerProps {
   user: PatronV5;
