+++ src/core/utils/types/loan-type.ts
@@ -13,3 +13,7 @@ export interface LoanType extends ListType {
 export function isLoanType(item: ListType): item is LoanType {
   return !!item.loanId;
 }
+
+export function loanId(loan: LoanType): string {
+  return String(loan.loanId || loan.identifier);
+}
